# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SERMACROPS EDI System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export interface EdiProcessResult {
  success: boolean;
  transactionId?: number;
  transactionType?: string;
  partnerId?: string;
  message: string;
  responseEdi?: string;
}

export interface EdiLineItem {
  lineNumber: number;
  productId: string;
  description?: string;
  quantity: number;
  unitPrice?: number;
  uom?: string;
}

export interface SimulateEdiRequest {
  partnerId: string;
  referenceNumber?: string;
  items?: EdiLineItem[];
}

export type TransactionDirection =
  (typeof TransactionDirection)[keyof typeof TransactionDirection];

export const TransactionDirection = {
  inbound: "inbound",
  outbound: "outbound",
} as const;

export type TransactionStatus =
  (typeof TransactionStatus)[keyof typeof TransactionStatus];

export const TransactionStatus = {
  pending: "pending",
  processed: "processed",
  acknowledged: "acknowledged",
  failed: "failed",
} as const;

export type TransactionIntegrityStatus =
  (typeof TransactionIntegrityStatus)[keyof typeof TransactionIntegrityStatus];

export const TransactionIntegrityStatus = {
  valid: "valid",
  invalid: "invalid",
  unknown: "unknown",
} as const;

export type TransactionParsedJson = { [key: string]: unknown };

export interface Transaction {
  id: number;
  transactionType: string;
  direction: TransactionDirection;
  partnerId: string;
  partnerName?: string;
  controlNumber: string;
  status: TransactionStatus;
  integrityStatus?: TransactionIntegrityStatus;
  rawEdi?: string;
  parsedJson?: TransactionParsedJson;
  errorMessage?: string;
  createdAt: string;
  updatedAt?: string;
}

export interface TransactionList {
  transactions: Transaction[];
  total: number;
}

export type PurchaseOrderDirection =
  (typeof PurchaseOrderDirection)[keyof typeof PurchaseOrderDirection];

export const PurchaseOrderDirection = {
  inbound: "inbound",
  outbound: "outbound",
} as const;

export type PurchaseOrderStatus =
  (typeof PurchaseOrderStatus)[keyof typeof PurchaseOrderStatus];

export const PurchaseOrderStatus = {
  pending: "pending",
  acknowledged: "acknowledged",
  shipped: "shipped",
  invoiced: "invoiced",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface PurchaseOrder {
  id: number;
  poNumber: string;
  direction: PurchaseOrderDirection;
  partnerId: string;
  partnerName?: string;
  status: PurchaseOrderStatus;
  totalAmount?: number;
  currency?: string;
  shipDate?: string;
  items?: EdiLineItem[];
  createdAt: string;
  updatedAt?: string;
}

export interface PurchaseOrderList {
  purchaseOrders: PurchaseOrder[];
  total: number;
}

export interface InventoryItem {
  id: number;
  productId: string;
  productName: string;
  category?: string;
  quantityOnHand: number;
  quantityReserved?: number;
  quantityAvailable: number;
  reorderPoint?: number;
  unitOfMeasure?: string;
  unitCost?: number;
  updatedAt: string;
}

export interface InventoryList {
  inventory: InventoryItem[];
  total: number;
}

export interface UpdateInventoryBody {
  quantityOnHand?: number;
  quantityReserved?: number;
}

export type PartnerType = (typeof PartnerType)[keyof typeof PartnerType];

export const PartnerType = {
  client: "client",
  supplier: "supplier",
  logistics: "logistics",
} as const;

export interface Partner {
  id: string;
  name: string;
  type: PartnerType;
  ediId: string;
  as2Id?: string;
  endpointUrl?: string;
  isActive: boolean;
}

export interface PartnerList {
  partners: Partner[];
}

export interface DashboardSummary {
  totalTransactions: number;
  pendingTransactions: number;
  pendingPurchaseOrders: number;
  totalInventoryItems: number;
  lowStockItems: number;
  activePartners: number;
  transactionsToday?: number;
  failedTransactions?: number;
}

export type TransactionStatsByTypeItem = {
  type: string;
  count: number;
  label: string;
};

export type TransactionStatsByStatusItem = {
  status: string;
  count: number;
};

export type TransactionStatsByPartnerItem = {
  partnerId: string;
  partnerName: string;
  count: number;
};

export interface TransactionStats {
  byType: TransactionStatsByTypeItem[];
  byStatus: TransactionStatsByStatusItem[];
  byPartner: TransactionStatsByPartnerItem[];
}

export type RecentActivityActivitiesItem = {
  id: number;
  transactionType: string;
  direction: string;
  partnerName: string;
  status: string;
  controlNumber?: string;
  createdAt: string;
};

export interface RecentActivity {
  activities: RecentActivityActivitiesItem[];
}

export type ListTransactionsParams = {
  limit?: number;
  offset?: number;
  partnerId?: string;
  transactionType?: string;
  status?: ListTransactionsStatus;
};

export type ListTransactionsStatus =
  (typeof ListTransactionsStatus)[keyof typeof ListTransactionsStatus];

export const ListTransactionsStatus = {
  pending: "pending",
  processed: "processed",
  acknowledged: "acknowledged",
  failed: "failed",
} as const;

export type ListPurchaseOrdersParams = {
  status?: ListPurchaseOrdersStatus;
  partnerId?: string;
};

export type ListPurchaseOrdersStatus =
  (typeof ListPurchaseOrdersStatus)[keyof typeof ListPurchaseOrdersStatus];

export const ListPurchaseOrdersStatus = {
  pending: "pending",
  acknowledged: "acknowledged",
  shipped: "shipped",
  invoiced: "invoiced",
  completed: "completed",
  cancelled: "cancelled",
} as const;