# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SERMACROPS EDI System API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  totalTransactions: number;
  pendingTransactions: number;
  pendingPurchaseOrders: number;
  totalInventoryItems: number;
  lowStockItems: number;
  activePartners: number;
  transactionsToday?: number;
  failedTransactions?: number;
}