# lib/api-zod/src/generated/types/ediLineItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SERMACROPS EDI System API
 * OpenAPI spec version: 0.1.0
 */

export interface EdiLineItem {
  lineNumber: number;
  productId: string;
  description?: string;
  quantity: number;
  unitPrice?: number;
  uom?: string;
}