# lib/api-zod/src/generated/types/ediProcessResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SERMACROPS EDI System API
 * OpenAPI spec version: 0.1.0
 */

export interface EdiProcessResult {
  success: boolean;
  transactionId?: number;
  transactionType?: string;
  partnerId?: string;
  message: string;
  responseEdi?: string;
}