# lib/api-zod/src/generated/types/partner.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SERMACROPS EDI System API
 * OpenAPI spec version: 0.1.0
 */
import type { PartnerType } from "./partnerType";

export interface Partner {
  id: string;
  name: string;
  type: PartnerType;
  ediId: string;
  as2Id?: string;
  endpointUrl?: string;
  isActive: boolean;
}