# lib/api-zod/src/generated/types/transactionStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SERMACROPS EDI System API
 * OpenAPI spec version: 0.1.0
 */
import type { TransactionStatsByPartnerItem } from "./transactionStatsByPartnerItem";
import type { TransactionStatsByStatusItem } from "./transactionStatsByStatusItem";
import type { TransactionStatsByTypeItem } from "./transactionStatsByTypeItem";

export interface TransactionStats {
  byType: TransactionStatsByTypeItem[];
  byStatus: TransactionStatsByStatusItem[];
  byPartner: TransactionStatsByPartnerItem[];
}